requirements.txt
flask
flask-cors
python-dotenv
groq
numpy 
tiktoken 
opencv-python